fix auth bug + pylint support

+++ app/utils/time.py
@@ -5,10 +5,9 @@ def get_utc() -> dt.datetime:
     return dt.datetime.now(tz=dt.timezone.utc)
 
 
-def get_utc_timestamp() -> int:
-    return int(get_utc().timestamp())
+def get_utc_timestamp() -> float:
+    return float(get_utc().timestamp())
 
 
 def timestamp_to_utc(timestamp: int) -> dt.datetime:
     return dt.datetime.fromtimestamp(timestamp, tz=dt.timezone.utc)
-
